Use `struct-field-info-list` instead
`struct-define` gets a list of field names by taking the list of ref
functions (point-x, point-y) and cutting out the first part of each
string (point-, point-), leaving a list of field names (x, y).

Since `struct-field-info-list` was added in Racket `base` 7.7.0.9,
it's now possible to query the list of field names directly, so I've
converted the code to use that instead.