Add support for #:prefix and #:separator kw args
Assume I have a struct (struct point (x y)), and an instance of the
struct (define p (point 2 3)).  Normally, struct-define allows me to
refer to the fields of p as simply x and y.

With this change, an optional #:prefix argument allows me to refer
to them instead as p.x and p.y.  If I specify the further argument
 #:separator ":" I can refer to them as, for example, p:x and p:y.